till inheritance.

+++ Controllers/DepartmentsController.cs
@@ -86,7 +86,7 @@ namespace HertfordshireUniversity.Controllers
             {
                 return NotFound();
             }
-            ViewData["InstructorID"] = new SelectList(_context.Instructors, 
+            ViewData["InstructorID"] = new SelectList(_context.Instructors,
                 "ID", "FullName", department.InstructorID);
             return View(department);
         }
@@ -148,12 +148,12 @@ namespace HertfordshireUniversity.Controllers
                     {
                         var databaseValues = (Department)databaseEntry.ToObject();
 
-                        if(databaseValues.Name != clientValues.Name)
+                        if (databaseValues.Name != clientValues.Name)
                         {
                             ModelState.AddModelError("Name", $"Current Value: {databaseValues.Name}");
                         }
 
-                        if(databaseValues.Budget != clientValues.Budget)
+                        if (databaseValues.Budget != clientValues.Budget)
                         {
                             ModelState.AddModelError("Budget", $"Current Value: {databaseValues.Budget}");
                         }
@@ -181,13 +181,13 @@ namespace HertfordshireUniversity.Controllers
                     }
                 }
             }
-            ViewData["InstructorID"] = new SelectList(_context.Instructors, 
+            ViewData["InstructorID"] = new SelectList(_context.Instructors,
                 "ID", "FullName", departmentToUpdate.InstructorID);
             return View(departmentToUpdate);
         }
 
         // GET: Departments/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? concurrencyError)
         {
             if (id == null)
             {
@@ -196,24 +196,53 @@ namespace HertfordshireUniversity.Controllers
 
             var department = await _context.Departments
                 .Include(d => d.Administrator)
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.DepartmentID == id);
             if (department == null)
             {
+                if (concurrencyError.GetValueOrDefault())
+                {
+                    return RedirectToAction(nameof(Index));
+                }
                 return NotFound();
+            }
+
+            if (concurrencyError.GetValueOrDefault())
+            {
+                ViewData["ConcurrencyErrorMessage"] = "The record you attempted to delete "
+                    + "was modified by another user after you got the original values. "
+                    + "The delete operation was canceled and the current values in the "
+                    + "database have been displayed. If you still want to delete this "
+                    + "record, click the Delete button again. Otherwise "
+                    + "click the Back to List hyperlink.";
             }
 
             return View(department);
         }
 
         // POST: Departments/Delete/5
-        [HttpPost, ActionName("Delete")]
+
+        [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> Delete(Department department)
         {
-            var department = await _context.Departments.FindAsync(id);
-            _context.Departments.Remove(department);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            try
+            {
+                if (await _context.Departments.AnyAsync(m => m.DepartmentID == department.DepartmentID))
+                {
+                    _context.Departments.Remove(department);
+                    await _context.SaveChangesAsync();
+                }
+                return RedirectToAction(nameof(Index));
+            }
+            catch(DbUpdateConcurrencyException)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new { concurrencyError = true, id = department.DepartmentID });
+            }
+
+            //var department = await _context.Departments.FindAsync(id);
+          
         }
 
         private bool DepartmentExists(int id)
